removed dummy enum. made genmode a byte where it's not an enum

+++ src/main/java/com/seibel/lod/core/builders/lodBuilding/LodBuilder.java
@@ -313,7 +313,7 @@ public class LodBuilder
 		
 		boolean hasCeiling = MC.getWrappedClientWorld().getDimensionType().hasCeiling();
 		boolean hasSkyLight = MC.getWrappedClientWorld().getDimensionType().hasSkyLight();
-		int generation = config.distanceGenerationMode.complexity;
+		byte generation = config.distanceGenerationMode.complexity;
 		int count = 0;
 		// FIXME: This yAbs is just messy!
 		int x = chunk.getMinX() + chunkSubPosX;
